count inversion
For a given array A[] we need to find the number of inversion in that array. For an array of ‘n’ elements an (i,j) pair is considered as an inversion if i<j for A[i]>A[j], where i and j are any two positions in the array A. I have  written a program to count the number of inversion in the array.